[parser] Skipping inner funcs: minor fixes.
1) Function recording conditions need to be consistent (this same condition is used above)

2) byte is not wide enough for storing the backing store size.

Bugs uncovered by the existing tests with the flag on.

BUG=v8:5516

+++ src/parsing/preparsed-scope-data.h
@@ -105,7 +105,7 @@ class PreParsedScopeData {
 
   // TODO(marja): Make the backing store more efficient once we know exactly
   // what data is needed.
-  std::vector<byte> backing_store_;
+  std::vector<int32_t> backing_store_;
 
   // Start pos -> FunctionData. Used for creating FunctionLiterals for skipped
   // functions (when they're actually skipped).
